refactor(web): extract useSavedPapers hook and SearchPaperList component from SearchPage
